editor/libeditor/text/tests/test_bug471722.html

<!DOCTYPE HTML>
<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Plaintext Editor Test code
   -
   - The Initial Developer of the Original Code is
   - Graeme McCutcheon <graememcc_firefox@graeme-online.co.uk>.
   - Portions created by the Initial Developer are Copyright (C) 2009
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=471722
-->

<head>
  <title>Test for Bug 471722</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
</head>

<body onload="doTest();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=471722">Mozilla Bug 471722</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script type="application/javascript">

      /** Test for Bug 471722 **/
    
      SimpleTest.waitForExplicitFinish();

      function doTest() {
        var t1 = $("t1");
        var editor = null;

        netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect"); 
        if (t1 instanceof Components.interfaces.nsIDOMNSEditableElement)
          editor = t1.editor;
        ok(editor, "able to get editor for the element");
        t1.focus();
        t1.select();

        try {

          // Cut the initial text in the textbox
          ok(editor.canCut(), "can cut text");
          editor.cut();
          is(t1.value, "", "initial text was removed");
 
          // So now we will have emptied the textfield
          // and the editor will have created a bogus node
          // Check the transaction is in the undo stack...
          var t1Enabled = {};
          var t1CanUndo = {};
          editor.canUndo(t1Enabled, t1CanUndo);
          ok(t1CanUndo.value, "undo is enabled");

          // Undo the cut
          editor.undo(1);
          is(t1.value, "minefield", "text reinserted");

          // So now, the cut should be in the redo stack,
          // so executing the redo will clear the text once again
          // and reinsert the bogus node that was removed after undo.
          // This will require the editor to figure out that we have a
          // bogus node again...
          var t1CanRedo = {};
          editor.canRedo(t1Enabled, t1CanRedo);
          ok(t1CanRedo.value, "redo is enabled");
          editor.redo(1);

          // Did the editor notice a bogus node reappeared?
          is(t1.value, "", "editor found bogus node");
       } catch (e) {
         ok(false, "test failed with error "+e);
       }
       SimpleTest.finish();
     }
   </script>
  </pre>

  <input type="text" value="minefield" id="t1" />
</body>
</html>
